Made Span abstract and added Enabled and Disabled (currently unused) subclasses
Step 2 of a refactoring to give Span two implementations, one for when tracing is disabled and one for when it's enabled.
