+++ telemetry-lib/src/main/java/com/yammer/telemetry/tracing/Span.java
@@ -3,66 +3,30 @@ package com.yammer.telemetry.tracing;
 import com.google.common.base.Optional;
 
 import java.math.BigInteger;
-import java.util.LinkedList;
 import java.util.List;
 
-/**
- * Start a new trace.
- * Start a new span within a trace.
- * Attach to a span within a trace.
- */
-public class Span implements AutoCloseable, SpanData {
+public abstract class Span implements AutoCloseable, SpanData {
+    protected final BigInteger traceId;
+    protected final Optional<BigInteger> parentSpanId;
+    protected final BigInteger spanId;
+    protected final String name;
+    protected final String host;
+    protected final Integer pid;
 
-    private final BigInteger traceId;
-    private final Optional<BigInteger> parentSpanId;
-    private final BigInteger spanId;
-    private final String name;
-    private final String host;
-    private final long startTime;
-    private final Integer pid;
-    private long duration;
-    private final TraceLevel traceLevel;
-    private final List<AnnotationData> annotations;
-
-    Span(BigInteger traceId, BigInteger spanId, Optional<BigInteger> parentSpanId, String name, long startTime, long startNanos, TraceLevel traceLevel) {
-        this.traceId = traceId;
+    protected Span(Optional<BigInteger> parentSpanId, BigInteger spanId, String name, BigInteger traceId) {
+        this.host = Annotations.getServiceAnnotations().getHost();
+        this.pid = Annotations.getServiceAnnotations().getPid();
         this.parentSpanId = parentSpanId;
         this.spanId = spanId;
         this.name = name;
-        this.host = Annotations.getServiceAnnotations().getHost();
-        this.pid = Annotations.getServiceAnnotations().getPid();
-        this.startTime = startTime;
-        this.duration = startNanos;
-        this.traceLevel = traceLevel;
-        this.annotations = new LinkedList<>();
+        this.traceId = traceId;
     }
 
-    public void addAnnotation(String name) {
-        annotations.add(new Annotation(SpanHelper.nowInNanoseconds(), name));
-    }
+    public abstract void addAnnotation(String name);
 
-    public void addAnnotation(String name, String message) {
-        annotations.add(new Annotation(SpanHelper.nowInNanoseconds(), name, message));
-    }
+    public abstract void addAnnotation(String name, String message);
 
-    public void end() {
-        duration = System.nanoTime() - duration;
-
-        // we need to ensure this span context is ended even if it's not being logged,
-        // otherwise we risk pollution of the context for subsequent operations.
-        SpanHelper.SpanContext context = SpanHelper.spanContext.get();
-        if (context != null) {
-            final Iterable<SpanSink> sinks = SpanSinkRegistry.getSpanSinks();
-            context.endSpan(this);
-            if (getTraceLevel() == TraceLevel.ON) {
-                for (SpanSink sink : sinks) {
-                    sink.record(this);
-                }
-            }
-        } else {
-            throw new IllegalStateException("Span.end() from a detached span.");
-        }
-    }
+    public abstract void end();
 
     @Override
     public void close() {
@@ -94,33 +58,12 @@ public class Span implements AutoCloseable, SpanData {
         return pid;
     }
 
-    public long getStartTime() {
-        return startTime;
-    }
+    public abstract long getStartTime();
 
-    public long getDuration() {
-        return duration;
-    }
+    public abstract long getDuration();
 
     @Override
-    public List<AnnotationData> getAnnotations() {
-        return annotations;
-    }
+    public abstract List<AnnotationData> getAnnotations();
 
-    TraceLevel getTraceLevel() {
-        return traceLevel;
-    }
-
-    @Override
-    public String toString() {
-        return "Span{" +
-                "traceId=" + traceId +
-                ", parentSpanId=" + parentSpanId +
-                ", spanId=" + spanId +
-                ", name='" + name + '\'' +
-                ", startTime=" + startTime +
-                ", duration=" + duration +
-                '}';
-    }
-
+    abstract TraceLevel getTraceLevel();
 }
